move extension enablement helper
