+++ shared/src/commandPalette/EmptyCommandList.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { SettingsCascadeOrError } from '../settings/settings'
-import { isErrorLike } from '../util/errors'
+import { onlyDefaultExtensionsAdded } from '../extensions/extensions'
 
 interface Props {
     settingsCascade?: SettingsCascadeOrError
@@ -59,29 +59,4 @@ export const EmptyCommandList: React.FunctionComponent<Props> = ({ settingsCasca
             <span className="empty-command-list__illustration">{puzzleIllustration}</span>
         </div>
     )
-}
-
-/**
- * Determines if only default extensions are added
- */
-function onlyDefaultExtensionsAdded(settings: SettingsCascadeOrError): boolean {
-    if (!isErrorLike(settings.subjects) && settings.subjects) {
-        const userSettings = settings.subjects.find(subject => subject.subject.__typename === 'User')
-        const defaultSettings = settings.subjects.find(subject => subject.subject.__typename === 'DefaultSettings')
-
-        if (userSettings && defaultSettings) {
-            const userExtensions = !isErrorLike(userSettings.settings) && userSettings.settings?.extensions
-            const defaultExtensions = !isErrorLike(defaultSettings.settings) && defaultSettings.settings?.extensions
-
-            if (userExtensions && defaultExtensions) {
-                for (const key of Object.keys(userExtensions)) {
-                    if (!(key in defaultExtensions)) {
-                        return false
-                    }
-                }
-            }
-        }
-    }
-
-    return true
 }
